Fixed bug in compiler!
There’s a bug in the compiler! It’s a hold-over from the source
material. Sample program that breaks it at present:

    l := 5; if !l >= 5 then (l := 0; true) else (false)

which roughly gets transformed to

    l := 5; if !l >= 5 then l := 0 else skip; if !l >= 5 then true else
false

which returns false, not the correct true value.

This fixes it by writing the value of the condition to a new location
and dereferencing when required, ensuring correctness.